[TAN-8560] Extract the custom page's pre-builder sections into a component
